refactor(build): move Go launcher build to dedicated script
