+++ scripts/copy-artifacts.js
@@ -3,7 +3,6 @@
 
 const fs = require("fs");
 const path = require("path");
-const { execFileSync } = require("child_process");
 
 const root = process.cwd();
 
@@ -32,21 +31,4 @@ const executables = [
 
 for (const p of executables) {
   if (fs.existsSync(p) && process.platform !== "win32") fs.chmodSync(p, 0o755);
-}
-
-// Cross-compile Windows launcher (requires Go)
-const launcherDir = path.join(root, "src", "host", "launcher");
-const outPath = path.join(root, "packages", "win32-x64", "tabctl-host.exe");
-fs.mkdirSync(path.dirname(outPath), { recursive: true });
-try {
-  execFileSync("go", ["build", "-o", outPath, "."], {
-    cwd: launcherDir,
-    stdio: "pipe",
-    env: { ...process.env, GOOS: "windows", GOARCH: "amd64", CGO_ENABLED: "0" },
-  });
-  console.log("Windows launcher built:", outPath);
-} catch (err) {
-  console.error("Failed to build Windows launcher. Is Go installed?");
-  console.error(err.stderr?.toString() || err.message);
-  process.exit(1);
 }
